move host_filter.rs to new cluster mod
