+++ scylla/src/cluster/cluster.rs
@@ -1,3 +1,4 @@
+use crate::cluster::host_filter::HostFilter;
 use crate::cluster::node::Node;
 use crate::connection::{Connection, PoolConfig, VerifiedKeyspaceName};
 /// Cluster manages up to date information and connections to database nodes
@@ -6,7 +7,6 @@ use crate::prepared_statement::TokenCalculationError;
 use crate::routing::{Shard, Token};
 use crate::session::TABLET_CHANNEL_SIZE;
 use crate::transport::errors::{BadQuery, NewSessionError, QueryError};
-use crate::transport::host_filter::HostFilter;
 use crate::transport::locator::ReplicaLocator;
 use crate::transport::partitioner::PartitionerName;
