+++ scylla/src/cluster/metadata.rs
@@ -1,3 +1,4 @@
+use crate::cluster::host_filter::HostFilter;
 use crate::cluster::node::resolve_contact_points;
 use crate::connection::{Connection, ConnectionConfig, NodeConnectionPool, PoolConfig, PoolSize};
 use crate::deserialize::DeserializeOwnedRow;
@@ -5,7 +6,6 @@ use crate::frame::response::event::Event;
 use crate::routing::Token;
 use crate::statement::query::Query;
 use crate::transport::errors::{DbError, NewSessionError, QueryError};
-use crate::transport::host_filter::HostFilter;
 use crate::transport::iterator::QueryPager;
 use crate::utils::parse::{ParseErrorCause, ParseResult, ParserState};
