+++ scylla/src/lib.rs
@@ -289,8 +289,8 @@ pub use transport::query_result::{QueryResult, QueryRowsResult};
 #[cfg(feature = "cloud")]
 pub use session::CloudSessionBuilder;
 
+pub use cluster::host_filter;
 pub use transport::execution_profile;
-pub use transport::host_filter;
 pub use transport::load_balancing;
 pub use transport::retry_policy;
 pub use transport::speculative_execution;
