+++ scylla/src/session/session_builder.rs
@@ -13,9 +13,9 @@ use crate::transport::errors::NewSessionError;
 #[cfg(feature = "cloud")]
 use crate::ExecutionProfile;
 
+use crate::cluster::host_filter::HostFilter;
 use crate::connection::PoolSize;
 use crate::statement::Consistency;
-use crate::transport::host_filter::HostFilter;
 use std::borrow::Borrow;
 use std::marker::PhantomData;
 use std::net::SocketAddr;
@@ -808,7 +808,7 @@ impl<K: SessionBuilderKind> GenericSessionBuilder<K> {
     /// should be opened to the node or not. The driver will also avoid
     /// those nodes when re-establishing the control connection.
     ///
-    /// See the [host filter](crate::transport::host_filter) module for a list
+    /// See the [host filter](crate::cluster::host_filter) module for a list
     /// of pre-defined filters. It is also possible to provide a custom filter
     /// by implementing the HostFilter trait.
     ///
@@ -819,7 +819,7 @@ impl<K: SessionBuilderKind> GenericSessionBuilder<K> {
     /// # use std::sync::Arc;
     /// # use scylla::{Session, SessionBuilder};
     /// # use scylla::connection::{AddressTranslator, TranslationError};
-    /// # use scylla::transport::host_filter::DcHostFilter;
+    /// # use scylla::cluster::host_filter::DcHostFilter;
     ///
     /// # async fn example() -> Result<(), Box<dyn std::error::Error>> {
     /// // The session will only connect to nodes from "my-local-dc"
